requirements.txt
appdirs==1.4.4
asgiref==3.2.10
astroid==2.4.2
boto==2.49.0
boto3==1.16.30
botocore==1.19.30
certifi==2020.6.20
colorama==0.4.3
distlib==0.3.1
dj-database-url==0.5.0
Django==3.1.2
django-cors-headers==3.5.0
django-crispy-forms==1.9.2
django-heroku==0.3.1
django-mathfilters==1.0.0
django-s3direct==1.1.8
django-silk==4.1.0
django-storages==1.10.1
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.5.0
filelock==3.0.12
gunicorn==20.0.4
isort==4.3.21
jmespath==0.10.0
lazy-object-proxy==1.4.3
lxml==4.6.2
markdown2==2.3.9
mccabe==0.6.1
pipenv==2020.8.13
psycopg2==2.8.6
pylint==2.5.3
pylint-django==2.3.0
pylint-plugin-utils==0.6
PyMySQL==0.10.1
PyJWT==1.7.1
python-dateutil==2.8.1
python-decouple==3.3
python-dotenv==0.15.0
pytz==2020.1
s3transfer==0.3.3
sentry_sdk==1.3.1
six==1.15.0
sqlparse==0.4.1
toml==0.10.1
urllib3==1.26.2
virtualenv==20.0.33
virtualenv-clone==0.5.4
whitenoise==5.2.0
wrapt==1.12.1